+++ test/cinder/filter/cinder_filter_test.exs
@@ -1,0 +1,4 @@
+defmodule Cinder.FilterTest do
+  use ExUnit.Case, async: true
+  doctest Cinder.Filter
+end